requirements.txt
requests==2.32.3
python-socketio==5.11.3
lark==1.1.9
python-dotenv==1.0.1
tenacity==9.0.0
websocket-client==1.8.0
flask==3.0.3
jinja2==3.1.4